Revert "修改项目名称，防止mac系统不兼容"
