+++ backend/nuget/readme.md
@@ -2,7 +2,7 @@
 
 开放源代码组织
 
-## XiHan.BasicApp - 曦寒基础应用
+## XiHan.App - 曦寒应用
 
 通用、全面的角色控制系统，基于曦寒框架和曦寒界面构建。
