+++ backend/scripts/service/XiHan.App.bat
@@ -1,0 +1,5 @@
+cd ..
+sc create XiHan.App binPath= %~dp0XiHan.App.exe start= auto 
+sc description XiHan.App "XiHan.App"
+Net Start XiHan.App
+pause
