+++ backend/scripts/service/XiHan.App.service
@@ -1,0 +1,18 @@
+[Unit]
+Description=XiHan
+
+[Service]
+WorkingDirectory=/home/XiHan.App/
+ExecStart=/usr/bin/dotnet /home/XiHan.App/XiHan.App.WebHost.dll --urls="http//*:9708" --ip="127.0.0.1"
+Restart=always
+RestartSec=10
+KillSignal=SIGINT
+SyslogIdentifier=XiHan.App
+User=root
+Environment=ASPNETCORE_ENVIRONMENT=Production
+Environment=DOTNET_PRINT_TELEMETRY_MESSAGE=false
+Environment=DOTNET_ROOT=/usr/share/dotnet/dotnet
+
+[Install]
+WantedBy=multi-user.target
+
