--- backend/scripts/service/XiHan.BasicApp.bat
@@ -1,5 +1,0 @@
-cd ..
-sc create XiHan.BasicApp binPath= %~dp0XiHan.BasicApp.exe start= auto 
-sc description XiHan.BasicApp "XiHan.BasicApp"
-Net Start XiHan.BasicApp
-pause
