--- backend/scripts/service/XiHan.BasicApp.service
@@ -1,18 +1,0 @@
-[Unit]
-Description=XiHan
-
-[Service]
-WorkingDirectory=/home/XiHan.BasicApp/
-ExecStart=/usr/bin/dotnet /home/XiHan.BasicApp/XiHan.BasicApp.WebHost.dll --urls="http//*:9708" --ip="127.0.0.1"
-Restart=always
-RestartSec=10
-KillSignal=SIGINT
-SyslogIdentifier=XiHan.BasicApp
-User=root
-Environment=ASPNETCORE_ENVIRONMENT=Production
-Environment=DOTNET_PRINT_TELEMETRY_MESSAGE=false
-Environment=DOTNET_ROOT=/usr/share/dotnet/dotnet
-
-[Install]
-WantedBy=multi-user.target
-
