+++ backend/src/main/XiHan.App.Application/XiHan.App.Application.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props"/>
+	<Import Project="..\..\..\props\version.props"/>
+
+	<PropertyGroup>
+		<Title>XiHan.App.Application</Title>
+		<AssemblyName>XiHan.App.Application</AssemblyName>
+		<PackageId>XiHan.App.Application</PackageId>
+		<Description>曦寒应用应用库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+</Project>
