+++ backend/src/main/XiHan.App.Domain.Shared/XiHan.App.Domain.Shared.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props"/>
+	<Import Project="..\..\..\props\version.props"/>
+
+	<PropertyGroup>
+		<Title>XiHan.App.Domain.Shared</Title>
+		<AssemblyName>XiHan.App.Domain.Shared</AssemblyName>
+		<PackageId>XiHan.App.Domain.Shared</PackageId>
+		<Description>曦寒应用领域共享库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+</Project>
