+++ backend/src/main/XiHan.App.Domain/XiHan.App.Domain.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props"/>
+	<Import Project="..\..\..\props\version.props"/>
+
+	<PropertyGroup>
+		<Title>XiHan.App.Domain</Title>
+		<AssemblyName>XiHan.App.Domain</AssemblyName>
+		<PackageId>XiHan.App.Domain</PackageId>
+		<Description>曦寒应用领域库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+</Project>
