+++ backend/src/main/XiHan.App.WebHost/XiHan.App.WebHost.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+	<Import Project="..\..\..\props\common.props"/>
+	<Import Project="..\..\..\props\version.props"/>
+
+	<PropertyGroup>
+		<Title>XiHan.App.WebHost</Title>
+		<AssemblyName>XiHan.App.WebHost</AssemblyName>
+		<PackageId>XiHan.App.WebHost</PackageId>
+		<Description>曦寒应用 Web 主机</Description>
+		<IsPackable>true</IsPackable>
+		<OutputType>Exe</OutputType>
+	</PropertyGroup>
+
+</Project>
