--- backend/src/main/XiHan.BasicApp.Application.Contracts/XiHan.BasicApp.Application.Contracts.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props" />
-	<Import Project="..\..\..\props\version.props" />
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.Application.Contracts</Title>
-		<AssemblyName>XiHan.BasicApp.Application.Contracts</AssemblyName>
-		<PackageId>XiHan.BasicApp.Application.Contracts</PackageId>
-		<Description>曦寒基础应用应用契约库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-</Project>
