--- backend/src/main/XiHan.BasicApp.Application/XiHan.BasicApp.Application.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props"/>
-	<Import Project="..\..\..\props\version.props"/>
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.Application</Title>
-		<AssemblyName>XiHan.BasicApp.Application</AssemblyName>
-		<PackageId>XiHan.BasicApp.Application</PackageId>
-		<Description>曦寒基础应用应用库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-</Project>
