--- backend/src/main/XiHan.BasicApp.Domain.Shared/XiHan.BasicApp.Domain.Shared.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props"/>
-	<Import Project="..\..\..\props\version.props"/>
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.Domain.Shared</Title>
-		<AssemblyName>XiHan.BasicApp.Domain.Shared</AssemblyName>
-		<PackageId>XiHan.BasicApp.Domain.Shared</PackageId>
-		<Description>曦寒基础应用领域共享库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-</Project>
