--- backend/src/main/XiHan.BasicApp.Domain/XiHan.BasicApp.Domain.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props"/>
-	<Import Project="..\..\..\props\version.props"/>
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.Domain</Title>
-		<AssemblyName>XiHan.BasicApp.Domain</AssemblyName>
-		<PackageId>XiHan.BasicApp.Domain</PackageId>
-		<Description>曦寒基础应用领域库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-</Project>
