--- backend/src/main/XiHan.BasicApp.WebHost/XiHan.BasicApp.WebHost.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-	<Import Project="..\..\..\props\common.props"/>
-	<Import Project="..\..\..\props\version.props"/>
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.WebHost</Title>
-		<AssemblyName>XiHan.BasicApp.WebHost</AssemblyName>
-		<PackageId>XiHan.BasicApp.WebHost</PackageId>
-		<Description>曦寒基础应用 Web 主机</Description>
-		<IsPackable>true</IsPackable>
-		<OutputType>Exe</OutputType>
-	</PropertyGroup>
-
-</Project>
