+++ backend/src/modules/XiHan.App.CodeGeneration.Application/XiHan.App.CodeGeneration.Application.csproj
@@ -1,0 +1,23 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props" />
+	<Import Project="..\..\..\props\version.props" />
+
+	<PropertyGroup>
+		<Title>XiHan.App.CodeGeneration.Application</Title>
+		<AssemblyName>XiHan.App.CodeGeneration.Application</AssemblyName>
+		<PackageId>XiHan.App.CodeGeneration.Application</PackageId>
+		<Description>曦寒应用代码生成应用库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="XiHan.Framework.CodeGeneration" Version="0.8.2" />
+		<PackageReference Include="XiHan.Framework.Ddd.Application" Version="0.8.2" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\main\XiHan.App.Application.Contracts\XiHan.App.Application.Contracts.csproj" />
+	</ItemGroup>
+
+</Project>
