+++ backend/src/modules/XiHan.App.CodeGeneration.Domain/XiHan.App.CodeGeneration.Domain.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props" />
+	<Import Project="..\..\..\props\version.props" />
+
+	<PropertyGroup>
+		<Title>XiHan.App.CodeGeneration.Domain</Title>
+		<AssemblyName>XiHan.App.CodeGeneration.Domain</AssemblyName>
+		<PackageId>XiHan.App.CodeGeneration.Domain</PackageId>
+		<Description>曦寒应用代码生成领域库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="XiHan.Framework.Ddd.Domain" Version="0.8.2" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\XiHan.App.CodeGeneration.Domain.Shared\XiHan.App.CodeGeneration.Domain.Shared.csproj" />
+	</ItemGroup>
+
+</Project>
