+++ backend/src/modules/XiHan.App.Rbac.Application.Contracts/XiHan.App.Rbac.Application.Contracts.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props" />
+	<Import Project="..\..\..\props\version.props" />
+
+	<PropertyGroup>
+		<Title>XiHan.App.Rbac.Application.Contracts</Title>
+		<AssemblyName>XiHan.App.Rbac.Application.Contracts</AssemblyName>
+		<PackageId>XiHan.App.Rbac.Application.Contracts</PackageId>
+		<Description>曦寒应用角色控制应用契约库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+	<ItemGroup>
+	  <PackageReference Include="XiHan.Framework.Ddd.Application.Contracts" Version="0.8.2" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\main\XiHan.App.Domain.Shared\XiHan.App.Domain.Shared.csproj" />
+	</ItemGroup>
+
+</Project>
