+++ backend/src/modules/XiHan.App.Rbac.Application/XiHan.App.Rbac.Application.csproj
@@ -1,0 +1,23 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props" />
+	<Import Project="..\..\..\props\version.props" />
+
+	<PropertyGroup>
+		<Title>XiHan.App.Rbac.Application</Title>
+		<AssemblyName>XiHan.App.Rbac.Application</AssemblyName>
+		<PackageId>XiHan.App.Rbac.Application</PackageId>
+		<Description>曦寒应用角色控制应用库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+	<ItemGroup>
+	  <PackageReference Include="XiHan.Framework.Ddd.Application" Version="0.8.2" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\main\XiHan.App.Application.Contracts\XiHan.App.Application.Contracts.csproj" />
+	  <ProjectReference Include="..\..\main\XiHan.App.Domain\XiHan.App.Domain.csproj" />
+	</ItemGroup>
+
+</Project>
