+++ backend/src/modules/XiHan.App.Rbac.Domain/XiHan.App.Rbac.Domain.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<Import Project="..\..\..\props\common.props" />
+	<Import Project="..\..\..\props\version.props" />
+
+	<PropertyGroup>
+		<Title>XiHan.App.Rbac.Domain</Title>
+		<AssemblyName>XiHan.App.Rbac.Domain</AssemblyName>
+		<PackageId>XiHan.App.Rbac.Domain</PackageId>
+		<Description>曦寒应用角色控制领域库</Description>
+		<OutputType>Library</OutputType>
+	</PropertyGroup>
+
+	<ItemGroup>
+	  <PackageReference Include="XiHan.Framework.Ddd.Domain" Version="0.8.2" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\main\XiHan.App.Domain.Shared\XiHan.App.Domain.Shared.csproj" />
+	</ItemGroup>
+
+</Project>
