--- backend/src/modules/XiHan.BasicApp.CodeGeneration.Application.Contracts/XiHan.BasicApp.CodeGeneration.Application.Contracts.csproj
@@ -1,22 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props" />
-	<Import Project="..\..\..\props\version.props" />
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.CodeGeneration.Application.Contracts</Title>
-		<AssemblyName>XiHan.BasicApp.CodeGeneration.Application.Contracts</AssemblyName>
-		<PackageId>XiHan.BasicApp.CodeGeneration.Application.Contracts</PackageId>
-		<Description>曦寒基础应用代码生成应用契约库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="XiHan.Framework.Ddd.Application.Contracts" Version="0.8.2" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\XiHan.BasicApp.CodeGeneration.Domain.Shared\XiHan.BasicApp.CodeGeneration.Domain.Shared.csproj" />
-	</ItemGroup>
-
-</Project>
