--- backend/src/modules/XiHan.BasicApp.CodeGeneration.Domain.Shared/XiHan.BasicApp.CodeGeneration.Domain.Shared.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props" />
-	<Import Project="..\..\..\props\version.props" />
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.CodeGeneration.Domain.Shared</Title>
-		<AssemblyName>XiHan.BasicApp.CodeGeneration.Domain.Shared</AssemblyName>
-		<PackageId>XiHan.BasicApp.CodeGeneration.Domain.Shared</PackageId>
-		<Description>曦寒基础应用代码生成领域共享库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="XiHan.Framework.Ddd.Domain.Shared" Version="0.8.2" />
-	</ItemGroup>
-
-</Project>
