--- backend/src/modules/XiHan.BasicApp.Rbac.Application.Contracts/XiHan.BasicApp.Rbac.Application.Contracts.csproj
@@ -1,22 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props" />
-	<Import Project="..\..\..\props\version.props" />
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.Rbac.Application.Contracts</Title>
-		<AssemblyName>XiHan.BasicApp.Rbac.Application.Contracts</AssemblyName>
-		<PackageId>XiHan.BasicApp.Rbac.Application.Contracts</PackageId>
-		<Description>曦寒基础应用角色控制应用契约库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="XiHan.Framework.Ddd.Application.Contracts" Version="0.8.2" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\..\main\XiHan.BasicApp.Domain.Shared\XiHan.BasicApp.Domain.Shared.csproj" />
-	</ItemGroup>
-
-</Project>
