--- backend/src/modules/XiHan.BasicApp.Rbac.Application/XiHanRbacApplicationModule.cs
@@ -1,35 +1,0 @@
-﻿#region <<版权版本注释>>
-
-// ----------------------------------------------------------------
-// Copyright ©2024 ZhaiFanhua All Rights Reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-// FileName:XiHanRbacApplicationModule
-// Guid:9b39d543-6e3f-46b8-a288-40076def6e6a
-// Author:zhaifanhua
-// Email:me@zhaifanhua.com
-// CreateTime:2024/12/7 6:24:50
-// ----------------------------------------------------------------
-
-#endregion <<版权版本注释>>
-
-using XiHan.Framework.Core.Modularity;
-using XiHan.Framework.Ddd.Application;
-
-namespace XiHan.BasicApp.Rbac.Application;
-
-/// <summary>
-/// XiHanRbacApplicationModule
-/// </summary>
-[DependsOn(
-    typeof(XiHanDddApplicationModule)
-    )]
-public class XiHanRbacApplicationModule : XiHanModule
-{
-    /// <summary>
-    /// 服务配置
-    /// </summary>
-    /// <param name="context"></param>
-    public override void ConfigureServices(ServiceConfigurationContext context)
-    {
-    }
-}
