--- backend/src/modules/XiHan.BasicApp.Rbac.Domain.Shared/XiHan.BasicApp.Rbac.Domain.Shared.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<Import Project="..\..\..\props\common.props" />
-	<Import Project="..\..\..\props\version.props" />
-
-	<PropertyGroup>
-		<Title>XiHan.BasicApp.Rbac.Domain.Shared</Title>
-		<AssemblyName>XiHan.BasicApp.Rbac.Domain.Shared</AssemblyName>
-		<PackageId>XiHan.BasicApp.Rbac.Domain.Shared</PackageId>
-		<Description>曦寒基础应用角色控制领域共享库</Description>
-		<OutputType>Library</OutputType>
-	</PropertyGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="XiHan.Framework.Ddd.Domain.Shared" Version="0.8.2" />
-	</ItemGroup>
-
-</Project>
